Rename internal/pkg/csv -> internal/pkg/slicer
