+++ internal/pkg/slicer/columnsparser/parse.go
