+++ internal/pkg/slicer/manifest/manifest.go
