+++ internal/pkg/slicer/manifest/manifest_test.go
