+++ internal/pkg/slicer/rowsreader/fixtures/bad_header.csv
