+++ internal/pkg/slicer/rowsreader/fixtures/two_rows.csv
