+++ internal/pkg/slicer/slicedwriter/fixtures/empty_with_new_line/part0001
