+++ internal/pkg/slicer/slicedwriter/fixtures/escaping/part0001
