+++ internal/pkg/slicer/slicedwriter/fixtures/multiple_parts_bytes_mode/part0002
