+++ internal/pkg/slicer/slicedwriter/fixtures/multiple_parts_rows_mode/part0003
