+++ internal/pkg/slicer/slicedwriter/slice.go
