+++ internal/pkg/slicer/slicedwriter/writer.go
